refactor reducer into separate file
