+++ screens/AppNavigator/screens/TransactionsScreen/screens/TransactionDetailScreen.tsx
@@ -2,79 +2,46 @@ import * as React from 'react'
 import tailwind from 'tailwind-rn'
 import { translate } from '../../../../../translations'
 import { View, Linking } from 'react-native'
-import BigNumber from 'bignumber.js'
 import { Text } from '../../../../../components'
 import { StackScreenProps } from '@react-navigation/stack'
 import { TransactionsParamList } from '../TransactionsNavigator'
 import { Ionicons } from '@expo/vector-icons'
+import { activityToTxRow } from './reducer'
+
+// FIXME(@ivan-zynesis): envvar
+const EXPLORER_BASE_URL = 'https://explorer.defichain.io'
 
 type Props = StackScreenProps<TransactionsParamList, 'TransactionDetailScreen'>
 
 export function TransactionDetailScreen (props: Props): JSX.Element {
-  const { params } = props.route
-
-  let color: 'green'|'gray'
-  let desc = ''
-  const isPositive = params.activity.vin === undefined
-
-  // TODO(@ivan-zynesis): fix when other token transaction can be included
-  const TOKEN_NAME: { [key in number]: string } = {
-    0: 'DFI',
-    1: 'tBTC'
-  }
-
-  const tokenId = TOKEN_NAME[params.activity.tokenId as number]
-  let amount = new BigNumber(params.activity.value)
-
-  if (isPositive) {
-    color = 'green'
-    // TODO(@ivan-zynesis): Simplified, more complicated token transaction should have different desc
-    desc = 'Received'
-  } else {
-    color = 'gray'
-    desc = 'Sent'
-    amount = amount.negated()
-  }
-
-  // reduced result
-  const option = {
-    id: params.activity.id,
-    desc,
-    color,
-    amount: amount.toFixed(),
-    block: params.activity.block.height,
-    token: tokenId,
-    txid: params.activity.txid,
-    url: ''
-  }
+  const { activity } = props.route.params
+  const option = activityToTxRow(activity)
 
   const grayDivider = <View style={tailwind('bg-gray w-full h-4')} />
+
   const RenderRow = (lhs: string, rhs: string): JSX.Element => {
     return (
       <View>
         {grayDivider}
         <View style={tailwind('bg-white p-2 border-b border-gray-200 flex-row items-center w-full h-16')}>
-          <View style={tailwind('w-1/2 flex-initial')}>
-            <Text style={tailwind('font-medium flex-initial')}>
-              {lhs}
-            </Text>
+          <View style={tailwind('w-1/2 flex-1')}>
+            <Text style={tailwind('font-medium')}>{lhs}</Text>
           </View>
-          <View style={tailwind('w-1/2 flex-end')}>
-            <Text style={tailwind('font-medium flex-end text-right text-gray-500')}>
-              {rhs}
-            </Text>
+          <View style={tailwind('w-1/2 flex-1')}>
+            <Text style={tailwind('font-medium text-right text-gray-600')}>{rhs}</Text>
           </View>
         </View>
       </View>
     )
   }
 
+  const url = explorerUrl(activity.txid)
   const onTxidUrlPressed = React.useCallback(async () => {
-    const supported = await Linking.canOpenURL(option.url)
+    const supported = await Linking.canOpenURL(url)
     if (supported) {
-      await Linking.openURL(option.url)
+      await Linking.openURL(url)
     }
-  }, [option.url])
+  }, [url])
 
   return (
     <View>
@@ -90,11 +57,15 @@ export function TransactionDetailScreen (props: Props): JSX.Element {
               {translate('screens/TransactionDetailScreen', option.txid)}
             </Text>
           </View>
-          <View style={tailwind('w-8 flex-grow-0 justify-center flex-end')}>
+          <View style={tailwind('w-8 flex-grow-0 justify-center')}>
             <Ionicons name='open-outline' size={24} color='pink' onPress={onTxidUrlPressed} />
           </View>
         </View>
       </View>
     </View>
   )
+}
+
+function explorerUrl (txid: string): string {
+  return `${EXPLORER_BASE_URL}/#/DFI/tx/${txid}`
 }
